lvl5: Supports running exe files as child proc
The shell supports running executable files as commands within child processes.

If the first piece of the input looks like a path (starts with . or /) a child
process is created and the command specified by the first argument is executed
within the child using the provided arguments similar to the exec builtin.

If executing the command fails, the child process prints an error message
including a description of the errno.

The shell waits for the child to finish running before printing the next prompt.

NOTE: execvp was removed and is not used!

Signed-off-by: yakir-david <yakir-david@spring2024-iit.kdlp.underground.software>