lvl6: Checks whether the file exists in PATH
In the case that the user types something that isn't a path or a builtin, before printing the
unrecognized command error, the shell checks whether a file with that name exists in each of
the directories listed in the PATH environment variable in order.

If a file, with that name is found, the search can stop and that file is executed with arguments
in a child process.

Only if no file is found in any of the directories, should the unrecognized command error be printed.

Signed-off-by: yakir-david <yakir-david@spring2024-iit.kdlp.underground.software>