lvl7: Home Directory Substitution
Before processing the entered commands, the shell performs home directory
substitution on the pieces (command name or arguments) that start with a ~.

The shell determines a username string by taking a substring of the piece
after the ~ until the end of the string or the first /, whichever comes first.

If the username string is empty, then the ~ is replaced with the value of
the HOME environment variable.

If the username string is not empty, the shell attempts to locate the the
user with that username and, if successful, replaces the ~ and the
username substring with their home directory.

If getpwnam does not locate such a user, the shell leaves the piece unmodified.

Signed-off-by: yakir-david <yakir-david@spring2024-iit.kdlp.underground.software>