lvl8: Shell supports in\out-put Redirection
As the shell is processing the commands and arguments
if it finds a < or >, it skips any whitespace characters
and attempts to treat the next piece of the input as a
filename for redirection. If there is no filename before
the end of the string, an error message is printed.

A command can have more than one redirection
(even more than one of the same type). If there are
multiple of the same redirections, then the right-most
one takes precedence.

A redirection beginning with < causes the shell to open
the corresponding file and replace the stdin of the
child process with the file descriptor of the open file.

A redirection beginning with > behaves similarly, but it
replaces the stdout of the process.

Signed-off-by: yakir-david <yakir-david@spring2024-iit.kdlp.underground.software>