lvl9: Shell supports Pipe
The shell supports the | (pipe) operator to chain
multiple commands and their inputs and outputs.

A command can have more than one redirection
(even more than one of the same type).
If there are multiple of the same redirections
then the right-most one takes precedence.

A redirection beginning with < causes the shell
to open the corresponding file and replace the
stdin of the child process with the file
descriptor of the open file

A redirection beginning with > behaves similarly
but it replaces the stdout of the process.

Signed-off-by: yakir-david <yakir-david@spring2024-iit.kdlp.underground.software>